pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "dinary"
dynamic = [ "version",]
description = "Server for [Dinary - your dinar diary](https://github.com/andgineer/dinary). Track expenses, scan receipts, analyze spending with AI"
keywords = [ "expenses", "budget", "receipts", ]
classifiers = [ "Programming Language :: Python :: 3", "License :: OSI Approved :: MIT License", "Operating System :: OS Independent",]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "gspread>=6.1.0",
    "google-auth>=2.38.0",
    "httpx>=0.28.0",
    "sr-invoice-parser>=1.0.3",
    "pydantic-settings>=2.8.0",
    "yoyo-migrations>=9.0.0",
    "holidays>=0.94",
    "tenacity>=9.1.4",
]
[[project.authors]]
name = "Andrey Sorokin"
email = "andrey@sorokin.engineer"

[project.license]
file = "LICENSE.txt"

[project.urls]
Homepage = "https://andgineer.github.io/dinary/"
Documentation = "https://andgineer.github.io/dinary/"

[tool.setuptools]
packages = "find:"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.manifest]
include = [ "README.md", "LICENSE.txt",]

[tool.ruff]
line-length = 99

[tool.hatch.version]
path = "src/dinary/__about__.py"

[dependency-groups]
# PEP 735 dev dependency group. Previously this lived under
# ``[tool.uv].dev-dependencies``; uv has since deprecated that in
# favour of the standard ``dependency-groups`` table and emits a
# noisy warning on every ``uv run`` until it is migrated.
dev = [
    "allure-pytest>=2.14.2",
    "invoke>=2.2.0",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-material>=9.6.13",
    "mkdocstrings[python]>=0.29.1",
    "pre-commit>=4.2.0",
    "pymdown-extensions>=10.15",
    "pyrefly>=0.50.0",
    "pytest-cov>=6.1.1",
    "pytest>=8.3.5",
    "anyio[trio]>=4.9.0",
    "rich>=15.0.0",
]

[project.scripts]
dinary = "dinary.main:main"
